+++ modules/client-common/src/context-core.ts
@@ -6,6 +6,7 @@ import {
 } from "@ethersproject/providers";
 import {
   InvalidAddressError,
+  InvalidGasEstimationFactorError,
   UnsupportedNetworkError,
   UnsupportedProtocolError,
 } from "@aragon/sdk-common";
@@ -335,9 +336,7 @@ export abstract class ContextCore {
     gasFeeEstimationFactor: number,
   ): number {
     if (gasFeeEstimationFactor < 0 || gasFeeEstimationFactor > 1) {
-      throw new Error(
-        "Gas estimation factor value should be a number between 0 and 1",
-      );
+      throw new InvalidGasEstimationFactorError();
     }
     return gasFeeEstimationFactor;
   }