+++ modules/client/src/addresslistVoting/internal/client/estimation.ts
@@ -5,7 +5,11 @@ import {
 } from "../../../client-common";
 import { IAddresslistVotingClientEstimation } from "../interfaces";
 import { toUtf8Bytes } from "@ethersproject/strings";
-import { boolArrayToBitmap, decodeProposalId } from "@aragon/sdk-common";
+import {
+  boolArrayToBitmap,
+  decodeProposalId,
+  SizeMismatchError,
+} from "@aragon/sdk-common";
 import { ClientCore, GasFeeEstimation } from "@aragon/sdk-client-common";
 
 /**
@@ -34,9 +38,7 @@ export class AddresslistVotingClientEstimation extends ClientCore
       params.failSafeActions?.length &&
       params.failSafeActions.length !== params.actions?.length
     ) {
-      throw new Error(
-        "Size mismatch: actions and failSafeActions should match",
-      );
+      throw new SizeMismatchError();
     }
     const allowFailureMap = boolArrayToBitmap(params.failSafeActions);