+++ modules/client/src/tokenVoting/internal/client/methods.ts
@@ -14,6 +14,7 @@ import {
   NoProviderError,
   ProposalCreationError,
   resolveIpfsCid,
+  SizeMismatchError,
   UnsupportedNetworkError,
 } from "@aragon/sdk-common";
 import {
@@ -122,9 +123,7 @@ export class TokenVotingClientMethods extends ClientCore
       params.failSafeActions?.length &&
       params.failSafeActions.length !== params.actions?.length
     ) {
-      throw new Error(
-        "Size mismatch: actions and failSafeActions should match",
-      );
+      throw new SizeMismatchError();
     }
     const allowFailureMap = boolArrayToBitmap(params.failSafeActions);